--HG--
branch : issue/HAST-150

+++ Hast.Abstractions/Hast.Transformer.Abstractions/SimpleMemory/SimpleMemoryDeviceManifestExtensions.cs
@@ -4,7 +4,7 @@ namespace Hast.Transformer.Abstractions.SimpleMemory
 {
     public static class SimpleMemoryDeviceManifestExtensions
     {
-        public static uint GetAvailableSimpleMemoryCellCount(this IDeviceManifest deviceManifest) =>
+        public static ulong GetAvailableSimpleMemoryCellCount(this IDeviceManifest deviceManifest) =>
             deviceManifest.AvailableMemoryBytes / SimpleMemory.MemoryCellSizeBytes;
     }
 }
